+++ app/services/update_transaction.rb
@@ -1,25 +1,36 @@
 class UpdateTransaction
-  attr_accessor :transaction
-
-  def initialize(transaction:)
-    self.transaction = transaction
+  def initialize(transaction:, user:, report:)
+    @transaction = transaction
+    @user = user
+    @report = report
   end
 
   def call(attributes: {})
     transaction.assign_attributes(attributes)
 
     convert_and_assign_value(transaction, attributes[:value])
+    changes = transaction.changes
+    success = transaction.save
 
-    result = if transaction.valid?
-      Result.new(transaction.save, transaction)
+    if success
+      HistoryRecorder
+        .new(user: user)
+        .call(
+          changes: changes,
+          reference: "Update to Transaction",
+          activity: transaction.parent_activity,
+          trackable: transaction,
+          report: report
+        )
+      Result.new(true, transaction)
     else
       Result.new(false, transaction)
     end
-
-    result
   end
 
   private
+
+  attr_reader :transaction, :user, :report
 
   def convert_and_assign_value(transaction, value)
     transaction.value = ConvertFinancialValue.new.convert(value.to_s)