+++ spec/services/history_recorder_spec.rb
@@ -25,14 +25,22 @@ RSpec.describe HistoryRecorder do
     let(:user) { double("user") }
     let(:reference) { double("reference") }
     let(:activity) { double("activity") }
+    let(:trackable) { double("trackable") }
     let(:report) { double("report") }
 
     it "creates a HistoricalEvent for each change provided" do
-      recorder.call(changes: changes, activity: activity, reference: reference, report: report)
+      recorder.call(
+        changes: changes,
+        activity: activity,
+        trackable: trackable,
+        reference: reference,
+        report: report
+      )
 
       expect(HistoricalEvent).to have_received(:create).with(
         user: user,
         activity: activity,
+        trackable: trackable,
         report: report,
         reference: reference,
         value_changed: "title",
@@ -43,6 +51,7 @@ RSpec.describe HistoryRecorder do
       expect(HistoricalEvent).to have_received(:create).with(
         user: user,
         activity: activity,
+        trackable: trackable,
         report: report,
         reference: reference,
         value_changed: "description",
@@ -60,7 +69,13 @@ RSpec.describe HistoryRecorder do
       end
 
       it "does NOT create a HistoricalEvent for that particular property" do
-        recorder.call(changes: changes, activity: activity, reference: reference, report: report)
+        recorder.call(
+          changes: changes,
+          activity: activity,
+          trackable: trackable,
+          reference: reference,
+          report: report
+        )
 
         expect(HistoricalEvent).not_to have_received(:create).with(
           hash_including(
@@ -70,11 +85,18 @@ RSpec.describe HistoryRecorder do
       end
 
       it "does create a HistoricalEvent for other properties in the batch" do
-        recorder.call(changes: changes, activity: activity, reference: reference, report: report)
+        recorder.call(
+          changes: changes,
+          activity: activity,
+          trackable: trackable,
+          reference: reference,
+          report: report
+        )
 
         expect(HistoricalEvent).to have_received(:create).with(
           user: user,
           activity: activity,
+          trackable: trackable,
           report: report,
           reference: reference,
           value_changed: "objectives",